wasi-http: Allow embedder to manage outgoing connections
This is a backport of https://github.com/bytecodealliance/wasmtime/pull/7288,
minus the tests.  The test suite has been refactored since the release-14.0.0
branch was created, so backporting the tests would be complicated and not worth
the effort.
